Codechange: expose saveload.h less to the rest of the game
