+++ src/cheat_gui.cpp
@@ -13,7 +13,6 @@
 #include "company_base.h"
 #include "company_func.h"
 #include "currency_func.h"
-#include "saveload/saveload.h"
 #include "vehicle_base.h"
 #include "textbuf_gui.h"
 #include "window_gui.h"
