+++ src/console_cmds.cpp
@@ -13,7 +13,7 @@
 #include "debug.h"
 #include "engine_func.h"
 #include "landscape.h"
-#include "saveload/saveload.h"
+#include "saveload/saveload_func.h"
 #include "network/core/network_game_info.h"
 #include "network/network.h"
 #include "network/network_func.h"
